create basic lookup module
