+++ optimism/src/mips/witness.rs
@@ -1,13 +1,10 @@
-use crate::cannon::{Page, State};
-use crate::keccak::lookups::Lookups;
-use crate::keccak::ArithOps;
-use crate::mips::interpreter::{Lookup, LookupTable};
 use crate::{
     cannon::{
-        Hint, Meta, Start, StepFrequency, VmConfiguration, PAGE_ADDRESS_MASK, PAGE_ADDRESS_SIZE,
-        PAGE_SIZE,
+        Hint, Meta, Page, Start, State, StepFrequency, VmConfiguration, PAGE_ADDRESS_MASK,
+        PAGE_ADDRESS_SIZE, PAGE_SIZE,
     },
-    keccak::environment::KeccakEnv,
+    keccak::{environment::KeccakEnv, ArithOps},
+    lookup::{Lookup, LookupTable, Lookups},
     mips::{
         column::Column,
         interpreter::{
@@ -135,7 +132,7 @@ impl<Fp: Field> InterpreterEnv for Env<Fp> {
         }
     }
 
-    fn add_lookup(&mut self, _lookup: interpreter::Lookup<Self::Variable>) {
+    fn add_lookup(&mut self, _lookup: Lookup<Self::Variable>) {
         // FIXME: Track the lookup values in the environment.
     }
